+++ pumaguard/server.py
@@ -394,6 +394,24 @@ def main(options: argparse.Namespace, presets: Preset):
         webui.add_image_directory(folder)
         logger.info("Watching folder: %s", folder)
 
+    # Also expose classified result folders in the UI (browse-only)
+    try:
+        Path(presets.classified_puma_dir).mkdir(parents=True, exist_ok=True)
+        Path(presets.classified_other_dir).mkdir(parents=True, exist_ok=True)
+        Path(presets.intermediate_dir).mkdir(parents=True, exist_ok=True)
+    except OSError as exc:  # pragma: no cover
+        logger.error("Could not ensure classified folders exist: %s", exc)
+
+    webui.add_image_directory(presets.classified_puma_dir)
+    webui.add_image_directory(presets.classified_other_dir)
+    webui.add_image_directory(presets.intermediate_dir)
+    logger.info(
+        "Classified browsing enabled for: %s, %s; intermediate: %s",
+        presets.classified_puma_dir,
+        presets.classified_other_dir,
+        presets.intermediate_dir,
+    )
+
     manager.start_all()
 
     lock = acquire_lock()
